requirements.txt
tensorflow==2.10.0
keras==2.10.0
typing_extensions==4.3.0
fastapi==0.85.0
uvicorn==0.18.3
python-multipart==0.0.5
pillow==9.2.0
matplotlib==3.6.0
numpy==1.23.3
requests==2.28.1

# Base YOLOV5 --------------------------------
opencv-python>=4.1.1
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0  # see https://pytorch.org/get-started/locally/ (recommended)
torchvision>=0.8.1
tqdm>=4.64.0

# Logging -------------------------------------
tensorboard>=2.4.1

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
